move spec getter out of ducks
