+++ web/src/js/__tests__/ducks/modes/utilsSpec.tsx
@@ -1,11 +1,11 @@
 import fetchMock, { enableFetchMocks } from "jest-fetch-mock";
 import { TStore } from "../tutils";
 import {
-    includeListenAddress,
     isActiveMode,
     parseSpec,
     updateMode,
 } from "../../../ducks/modes/utils";
+import { includeListenAddress } from "../../../modes";
 
 enableFetchMocks();
