+++ web/src/js/modes/regular.ts
@@ -1,0 +1,7 @@
+import { includeListenAddress, ModeState } from ".";
+
+export interface RegularState extends ModeState {}
+
+export const getSpec = (m: RegularState): string => {
+    return includeListenAddress("regular", m);
+};
