+++ web/src/js/modes/wireguard.ts
@@ -1,0 +1,9 @@
+import { includeListenAddress, ModeState } from ".";
+
+export interface WireguardState extends ModeState {
+    file_path?: string;
+}
+
+export const getSpec = (s: WireguardState): string => {
+    return includeListenAddress("wireguard", s);
+};
